+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableArray.kt
@@ -217,13 +217,24 @@ public value class ImmutableArray<out T> @PublishedApi internal constructor(
 
     /**
      * Leaves this immutable array as is and returns an ImmutableArray with all elements sorted
-     * according to the natural sort order of the value returned by specified [selector].
+     * according to the natural sort order of the value returned by the [selector].
      *
      * The sort is _stable_ so equal elements preserve their order relative to each other after
      * sorting.
      */
     public inline fun <R : Comparable<R>> sortedBy(crossinline selector: (element: T) -> R?):
             ImmutableArray<T> = sortedWith(compareBy(selector))
+
+    /**
+     * Leaves this immutable array as is and returns an ImmutableArray with all elements sorted
+     * according to the reverse natural sort order of the value returned by the [selector].
+     *
+     * The sort is _stable_ so equal elements preserve their order relative to each other after
+     * sorting.
+     */
+    public inline fun <R : Comparable<R>> sortedByDescending(crossinline
+            selector: (element: T) -> R?): ImmutableArray<T> =
+            sortedWith(compareByDescending(selector))
 
     /**
      * Leaves this immutable array as is and returns an [ImmutableArray] with all elements sorted