+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableDoubleArray.kt
@@ -219,13 +219,24 @@ public value class ImmutableDoubleArray @PublishedApi internal constructor(
 
     /**
      * Leaves this immutable array as is and returns an ImmutableDoubleArray with all elements
-     * sorted according to the natural sort order of the value returned by specified [selector].
+     * sorted according to the natural sort order of the value returned by the [selector].
      *
      * The sort is _stable_ so equal elements preserve their order relative to each other after
      * sorting.
      */
     public inline fun <R : Comparable<R>> sortedBy(crossinline selector: (element: Double) -> R?):
             ImmutableDoubleArray = sortedWith(compareBy(selector))
+
+    /**
+     * Leaves this immutable array as is and returns an ImmutableDoubleArray with all elements
+     * sorted according to the reverse natural sort order of the value returned by the [selector].
+     *
+     * The sort is _stable_ so equal elements preserve their order relative to each other after
+     * sorting.
+     */
+    public inline fun <R : Comparable<R>> sortedByDescending(crossinline
+            selector: (element: Double) -> R?): ImmutableDoubleArray =
+            sortedWith(compareByDescending(selector))
 
     /**
      * Leaves this immutable array as is and returns an [ImmutableDoubleArray] with all elements